[Decomp][Player] Cohesion split: extract mail methods to PlayerMail.cpp
