+++ src/game/Object/Player.cpp
@@ -3300,129 +3300,6 @@ void Player::SendInitialSpells()
 }
 
 /**
- * @brief Removes a mail entry from the player's in-memory mailbox.
- *
- * @param id The message identifier to remove.
- */
-void Player::RemoveMail(uint32 id)
-{
-    for (PlayerMails::iterator itr = m_mail.begin(); itr != m_mail.end(); ++itr)
-    {
-        if ((*itr)->messageID == id)
-        {
-            // do not delete item, because Player::removeMail() is called when returning mail to sender.
-            m_mail.erase(itr);
-            return;
-        }
-    }
-}
-
-/**
- * @brief Sends the result of a mail operation to the client.
- *
- * @param mailId The mail message identifier.
- * @param mailAction The mail action that was processed.
- * @param mailError The result code for the action.
- * @param equipError The equipment error code used for equip-related failures.
- * @param item_guid The related item GUID low part.
- * @param item_count The related item count.
- */
-void Player::SendMailResult(uint32 mailId, MailResponseType mailAction, MailResponseResult mailError, uint32 equipError, uint32 item_guid, uint32 item_count)
-{
-    WorldPacket data(SMSG_SEND_MAIL_RESULT, (4 + 4 + 4 + (mailError == MAIL_ERR_EQUIP_ERROR ? 4 : (mailAction == MAIL_ITEM_TAKEN ? 4 + 4 : 0))));
-    data << (uint32) mailId;
-    data << (uint32) mailAction;
-    data << (uint32) mailError;
-    if (mailError == MAIL_ERR_EQUIP_ERROR)
-    {
-        data << (uint32) equipError;
-    }
-    else if (mailAction == MAIL_ITEM_TAKEN)
-    {
-        data << (uint32) item_guid;                         // item guid low?
-        data << (uint32) item_count;                        // item count?
-    }
-    GetSession()->SendPacket(&data);
-}
-
-/**
- * @brief Notifies the client that new mail is available.
- */
-void Player::SendNewMail()
-{
-    // deliver undelivered mail
-    WorldPacket data(SMSG_RECEIVED_MAIL, 4);
-    data << float(0.0f);
-    GetSession()->SendPacket(&data);
-}
-
-/**
- * @brief Recalculates unread mail count and the next pending delivery time.
- */
-void Player::UpdateNextMailTimeAndUnreads()
-{
-    // calculate next delivery time (min. from non-delivered mails
-    // and recalculate unReadMail
-    time_t cTime = time(NULL);
-    m_nextMailDelivereTime = 0;
-    unReadMails = 0;
-    for (PlayerMails::iterator itr = m_mail.begin(); itr != m_mail.end(); ++itr)
-    {
-        if ((*itr)->deliver_time > cTime)
-        {
-            if (!m_nextMailDelivereTime || m_nextMailDelivereTime > (*itr)->deliver_time)
-            {
-                m_nextMailDelivereTime = (*itr)->deliver_time;
-            }
-        }
-        else if (((*itr)->checked & MAIL_CHECK_MASK_READ) == 0)
-        {
-            ++unReadMails;
-        }
-    }
-}
-
-/**
- * @brief Tracks a newly scheduled mail delivery time.
- *
- * @param deliver_time The time when the mail should be delivered.
- */
-void Player::AddNewMailDeliverTime(time_t deliver_time)
-{
-    if (deliver_time <= time(NULL))                         // ready now
-    {
-        ++unReadMails;
-        SendNewMail();
-    }
-    else                                                    // not ready and no have ready mails
-    {
-        if (!m_nextMailDelivereTime || m_nextMailDelivereTime > deliver_time)
-        {
-            m_nextMailDelivereTime =  deliver_time;
-        }
-    }
-}
-
-
-/**
- * @brief Finds a mail entry by message identifier.
- *
- * @param id The message identifier to search for.
- * @return The matching mail entry, or null if none exists.
- */
-Mail* Player::GetMail(uint32 id)
-{
-    for (PlayerMails::iterator itr = m_mail.begin(); itr != m_mail.end(); ++itr)
-    {
-        if ((*itr)->messageID == id)
-        {
-            return (*itr);
-        }
-    }
-    return NULL;
-}
-
-/**
  * @brief Populates the create update mask for a target player.
  *
  * @param updateMask The update mask being built.
